refactor index.vue
